[avalanche] Fix ambiguous casts from size_t to UniValue
Summary: Introduced in D11696 and D11697

Test Plan:
  Build OSX

Reviewers: #bitcoin_abc, sdulfari, tobias_ruck

Reviewed By: #bitcoin_abc, tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D11720

+++ src/rpc/avalanche.cpp
@@ -757,8 +757,9 @@ static RPCHelpMan getavalancheinfo() {
                 network.pushKV("proof_count", proofCount);
                 network.pushKV("connected_proof_count", connectedProofCount);
                 network.pushKV("conflicting_proof_count",
-                               pm.getConflictingProofCount());
-                network.pushKV("orphan_proof_count", pm.getOrphanProofCount());
+                               uint64_t(pm.getConflictingProofCount()));
+                network.pushKV("orphan_proof_count",
+                               uint64_t(pm.getOrphanProofCount()));
                 network.pushKV("total_stake_amount", totalStakes);
                 network.pushKV("connected_stake_amount", connectedStakes);
